update remote access handlers & menu

+++ bin/menu-items-handlers.sh
@@ -458,23 +458,32 @@ show_app_logs() {
 
 start_tmate_session() {
   run_this_menu_item_handler_in_new_terminal_if_applicable || {
-    local start_pattern="# start mdm keys" end_pattern="# end mdm keys" tmate_socket
+    local start_pattern="# start mdm keys" end_pattern="# end mdm keys" auth_keys_file="$HOME/.ssh/authorized_keys" \
+      auth_keys_md5 auth_keys_to_add tmate_socket
+    if [[ "$mdm_tmate_authorized_keys_url" ]]; then
+      auth_keys_to_add="$(curl -sL "$mdm_tmate_authorized_keys_url")"
+      [[ "$auth_keys_to_add" =~ ssh\- ]] ||
+        error "Url $mdm_tmate_authorized_keys_url does not contain valid public ssh key(s)."
+      [[ ! -d "$HOME/.ssh/" ]] && mkdir "$HOME/.ssh/"
+      chmod 700 "$HOME/.ssh/"
+      [[ ! -f "$auth_keys_file" ]] && touch "$auth_keys_file"
+      chmod 600 "$auth_keys_file"
+      auth_keys_md5="$(md5sum "$auth_keys_file")"
+      perl -i.bak -0777 -pe "s/$start_pattern.*$end_pattern\r?\n//s" "$auth_keys_file" # rm old if exists
+      printf "%s\n%s\n%s\n" "$start_pattern" "$auth_keys_to_add" "$end_pattern" >> "$auth_keys_file" # add new
+      [[ "$auth_keys_md5" = "$(md5sum "$auth_keys_file")" ]] ||
+        msg_w_newlines "Successfully updated authorized keys."
+    else
+      warning_w_newlines "No authorized ssh keys set. (Please refer to the documentation.)
+If you continue, anyone with your unique url will be able to access to your system."
+      confirm_or_exit
+    fi
     tmate_socket="/tmp/tmate.$("$date_cmd" "+%s")"
-    [[ ! -d "$HOME/.ssh/" ]] && mkdir "$HOME/.ssh/"
-    chmod 700 "$HOME/.ssh/"
-    [[ ! -f "$HOME/.ssh/authorized_keys" ]] && touch "$HOME/.ssh/authorized_keys"
-    chmod 600 "$HOME/.ssh/authorized_keys"
-    perl -i.bak -0777 -pe "s/$start_pattern.*$end_pattern\r?\n//s" "$HOME/.ssh/authorized_keys"
-    { 
-      echo "$start_pattern"
-      curl -sL "$mdm_tmate_authorized_keys_url"
-      echo "$end_pattern"
-    } >> "$HOME/.ssh/authorized_keys"
     [[ "$(pgrep tmate)" ]] && { 
-      pkill tmate
-      sleep 5
+      pkill tmate # kill any existing session
+      sleep 3
     }
-    tmate -a "$HOME/.ssh/authorized_keys" -S "$tmate_socket" new-session -d
+    tmate -a "$auth_keys_file" -S "$tmate_socket" new-session -d
     tmate -S "$tmate_socket" wait tmate-ready
     ssh_url="$(tmate -S "$tmate_socket" display -p '#{tmate_ssh}')"
     msg_w_newlines "Provide this url to your remote collaborator. Access will end when they close their session or after a period of inactivity."
@@ -482,19 +491,55 @@ start_tmate_session() {
   }
 }
 
-toggle_remote_web_access() {
-  local remote_port local_port
-  remote_port="$(( $RANDOM + 20000 ))" # RANDOM is between 0-32767 (2^16 / 2 - 1)
-  local_port="$(docker ps -a --filter "name=varnish" --format "{{.Ports}}" | tr ',' '\n' | perl -ne "s/.*:(?=\d{5})// and s/-.*// and print")"
-  [[ ! "$local_port" =~ ^[0-9]+$ ]] && echo "Could not find valid local port" && exit 1
-  ssh -f \
-    -F /dev/null \
-    -o IdentitiesOnly=yes \
-    -o ExitOnForwardFailure=yes \
-    -o StrictHostKeyChecking=no \
-    -i "$HOME/pk" \
-    -NR "$remote_port":127.0.0.1:"$local_port" \
-    "$TUNNEL_SSH_URL"
+start_remote_web_access() {
+  run_this_menu_item_handler_in_new_terminal_if_applicable || {
+    if pgrep -f "ssh.*$mdm_tunnel_ssh_url"; then
+      :
+    fi
+    local remote_port local_port tmp_file hostname
+    remote_port="$(( $RANDOM + 20000 ))" # RANDOM is between 0-32767 (2^16 / 2 - 1)
+    local_port="$(docker ps -a --filter "name=varnish" --filter "label=com.docker.compose.project=$COMPOSE_PROJECT_NAME" --format "{{.Ports}}" | \
+      tr ',' '\n' | \
+      perl -ne "s/.*:(?=\d{5})// and s/-.*// and print"
+    )"
+    [[ ! "$local_port" =~ ^[0-9]+$ ]] && echo "Could not find valid local port" && exit 1
+    tmp_file="$(mktemp)"
+    get_github_file_contents "$mdm_tunnel_pk_url" > "$tmp_file"
+    ssh -f \
+      -F /dev/null \
+      -o IdentitiesOnly=yes \
+      -o ExitOnForwardFailure=yes \
+      -o StrictHostKeyChecking=no \
+      -i "$tmp_file" \
+      -NR "$remote_port":127.0.0.1:"$local_port" \
+      "$mdm_tunnel_ssh_url"
+    # TODO !DRY - similar to change_base_url
+    hostname="$remote_port.$mdm_tunnel_domain"
+    set_hostname_for_this_app "$hostname"
+    run_as_bash_cmds_in_app "
+      /app/bin/magento config:set web/unsecure/base_url https://$hostname/
+      /app/bin/magento config:set web/secure/base_url https://$hostname/
+      /app/bin/magento cache:flush
+    "
+    reload_rev_proxy
+    warm_cache
+    rm "$tmp_file"
+  }
+}
+
+stop_remote_access() {
+  run_this_menu_item_handler_in_new_terminal_if_applicable || {
+    if pkill tmate; then
+      msg_w_newlines "Succcessfully stopped 1 or more remote system access sessions."
+    else
+      msg_w_newlines "No active remote access sessions."
+    fi
+    if pkill -f "ssh.*$mdm_tunnel_ssh_url"; then
+      msg_w_newlines "Succcessfully stopped 1 or more remote web sessions."
+    else
+      msg_w_newlines "No active remote web sessions."
+    fi
+  }
 }
 
 show_errors_from_mdm_logs() {
